Generating the data for the view
